Provide invoke() and friends publicly in traits.h
This commit extracts the definitions of the C++17 backports of

invoke,
invoke_result,
invoke_result_t,
is_invokable,
is_invokable_r

from M. Park's variant header and makes them available in the
public traits.h header in namespace gul14.
